High Score Banner decoupled from distance banner
The High Score banner now maintains it's position in the world and is only reset when a new high-score overrides it.

Both the distance banner and high score banner use the ground as a reference of where they should be placed vertically to always point the correct distance to the ground.